+++ pcb/apple_keyboard_jis.csv
@@ -1,5 +1,5 @@
 1.5,1,1,1,1,1,1,1,1,1,1,1,1,1
-1,1,1,1,1,1,1,1,1,1,1,1,1,ISOEnter
-1.25,1,1,1,1,1,1,1,1,1,1,1,1,
+1,1,1,1,1,1,1,1,1,1,1,1,1,ISOEnter_Top
+1.25,1,1,1,1,1,1,1,1,1,1,1,1,ISOEnter_Bottom
 1.75,1,1,1,1,1,1,1,1,1,1,1,1.75,
 1,1,1.25,1.25,3.5,1.25,1.25,1,3,,,,
