+++ src/index.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
-import { store } from '../src/reducer';
 import App from './app';
+import { store } from './reducer';
 
 ReactDOM.render(
   <Provider store={store}>
